ref: externalized general create card into GeneralSection
The reason for this change is that this way the code for the general-information is clearly separated and thus easier to maintain.
