+++ src/app/checks/create/page.tsx
@@ -3,15 +3,12 @@ import { CreateCheckStoreProvider } from '@/components/check/create/CreateCheckP
 import Card from '@/components/Shared/Card'
 import PageHeading from '@/components/Shared/PageHeading'
 import insertKnowledgeCheck from '@/database/knowledgeCheck/insert'
+import GeneralSection from '@/src/app/checks/create/GeneralSection'
 import { SaveCreateCheckButton } from '@/src/app/checks/create/SaveCheckButton'
 import { Button } from '@/src/components/shadcn/button'
-import { Textarea } from '@/src/components/shadcn/textarea'
-import Input from '@/src/components/Shared/form/Input'
 import { getServerSession } from '@/src/lib/auth/server'
 import { getUUID } from '@/src/lib/Shared/getUUID'
 import { unauthorized } from 'next/navigation'
-import { ComponentType, InputHTMLAttributes } from 'react'
-import { twMerge } from 'tailwind-merge'
 
 export default async function CreateCheckPage() {
   const { user } = await getServerSession()
@@ -78,29 +75,7 @@ export default async function CreateCheckPage() {
     <CreateCheckStoreProvider>
       <PageHeading title='Create KnowledgeCheck' />
       <div className='columns-xl gap-12 space-y-12'>
-        <Card className='@container flex flex-col gap-8 p-3' disableHoverStyles>
-          <div className='header -m-3 flex flex-col rounded-t-md border-b border-neutral-400 bg-neutral-300 p-2 px-3 text-neutral-600 dark:border-neutral-500 dark:bg-neutral-700/60 dark:text-neutral-300'>
-            <div className='flex items-center justify-between'>
-              <h2 className=''>General Information</h2>
-            </div>
-          </div>
-          <div className='grid grid-cols-[auto_1fr] items-center gap-9 gap-x-7 p-2'>
-            <InputGroup label='Name' placeholder='Enter the name of your knowledge check' />
-            <InputGroup label='Description' className='min-h-20 resize-none' as={Textarea} placeholder='Describe the concept of your knowledge check using a few words.' />
-            <InputGroup
-              label='Deadline'
-              type='date'
-              defaultValue={new Date(Date.now())
-                .toLocaleDateString('de')
-                .split('.')
-                .reverse()
-                .map((el) => (el.length < 2 ? '0' + el : el))
-                .join('-')}
-              className='text-sm text-neutral-500 dark:text-neutral-400'
-            />
-            <InputGroup label='Administrators' />
-          </div>
-        </Card>
+        <GeneralSection />
         <Card disableHoverStyles className='break-inside-avoid'>
           <h2 className='text-lg'>Settings</h2>
           <div className='h-[500px]'></div>
@@ -116,23 +91,5 @@ export default async function CreateCheckPage() {
       </form>
       <div />
     </CreateCheckStoreProvider>
-  )
-}
-
-function InputGroup<E extends ComponentType>({ label, as, ...props }: { label: string; as?: E } & InputHTMLAttributes<HTMLInputElement>) {
-  const Element = as || Input
-
-  return (
-    <>
-      <label className='text-neutral-600 dark:text-neutral-400'>{label}</label>
-      <Element
-        placeholder='Enter some text'
-        {...props}
-        className={twMerge(
-          'rounded-md px-3 py-1.5 text-neutral-600 ring-1 ring-neutral-400 outline-none placeholder:text-neutral-400/90 hover:cursor-text hover:ring-neutral-500 focus:ring-[1.2px] focus:ring-neutral-700 dark:text-neutral-300/80 dark:ring-neutral-500 dark:placeholder:text-neutral-400/50 dark:hover:ring-neutral-300/60 dark:focus:ring-neutral-300/80',
-          props.className,
-        )}
-      />
-    </>
   )
 }
